# frontend/cinebook/src/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Globsyn Theatres API
  description: |-
    API for managing Globsyn Theatres online ticket booking system

    Contacts:
    - Adrita Chatterjee: adritachatterji@gmail.com
    - Anamitra Chandra:
    - Souraajit Samanta:
    - Srinjay Das Gupta: dasguptasrinjay2004@gmail.com, https://www.srinjaydg.in
  contact:
    name: Srinjay Das Gupta
    email: dasguptasrinjay2004@gmail.com
    url: https://www.srinjaydg.in
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Development server
paths:
...
  /ticket:
    get:
      summary: Retrieve a list of tickets from the database
      tags:
        - Tickets
      description: Get a list of all tickets
      responses:
        '200':
          description: A list of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketResponse'
    post:
      summary: Add a new ticket to the database
      tags:
        - Tickets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketRequest'
      responses:
        '201':
          description: The ticket was successfully added to the database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '500':
          description: Some server error
  /ticket/{id}:
    get:
      summary: Retrieve a ticket by ID
      tags:
        - Tickets
      parameters:
        - $ref: '#/components/parameters/ticketId'
      description: Retrieve a ticket from the database by its ID
      responses:
        '200':
          description: The ticket with the specified ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          description: The ticket with the specified ID was not found
    put:
      summary: Update a ticket by ID
      tags:
        - Tickets
      parameters:
        - $ref: '#/components/parameters/ticketId'
      description: Update a ticket by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketRequest'
      responses:
        '200':
          description: The ticket was successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          description: The ticket with the specified ID was not found
        '500':
          description: Some server error
    delete:
      summary: Delete a ticket by ID
      tags:
        - Tickets
      parameters:
        - $ref: '#/components/parameters/ticketId'
      description: Delete a ticket by ID
      responses:
        '200':
          description: The ticket was successfully deleted
        '404':
          description: The ticket with the specified ID was not found
        '500':
          description: Some server error
  /user:
    get:
      summary: Retrieve a list of users from the database
      tags:
        - Users
      description: Get a list of all users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
    post:
      summary: Add a new user to the database
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: The movie was successfully added to the database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '500':
          description: Some server error
  /user/{id}:
    get:
      summary: Retrieve a user by ID
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/userId'
      description: Retrieve a list of users from the database
      responses:
        '200':
          description: The user with the specified ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: The movie with the specified ID was not found
    put:
      summary: Update a user by ID
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/userId'
      description: Update an user by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: The user was successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: The user with the specified ID was not found
        '500':
          description: Some server error
    delete:
      summary: Delete an user by ID
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/userId'
      description: Delete a movie by ID
      responses:
        '200':
          description: The user was successfully deleted
        '404':
          description: The user with the specified ID was not found
        '500':
          description: Some server error
components:
  schemas:
...
    TicketRequest:
      type: object
      required:
        - name
        - date
        - time
        - movie
        - schedule
        - screen
        - seat
        - seatType
        - price
      properties:
        id:
          type: string
          description: Auto-generated unique identifier
        name:
          type: string
          description: Name of the ticket holder or event
        date:
          type: string
          format: date
          description: Date of the event
        time:
          type: string
          description: Time of the event in HH:mm format
        movie:
          type: string
          description: ID of the movie
        schedule:
          type: string
          description: ID of the schedule
        screen:
          type: number
          description: Screen number
        seat:
          type: array
          items:
            type: string
          description: List of seat numbers
        seatType:
          type: string
          description: Type of seat (e.g., regular, VIP)
        price:
          type: number
          description: Price of the ticket
      example:
        name: John Doe
        date: '2024-08-10'
        time: '19:30'
        movie: 60c72b2f9e7d3c001f1d1d6f
        schedule: 60c72b2f9e7d3c001f1d1d6f
        screen: 1
        seat:
          - A12
          - A13
        seatType: VIP
        price: 25
    TicketResponse:
      type: object
      required:
        - name
        - date
        - time
        - movie
        - schedule
        - screen
        - seat
        - seatType
        - price
      properties:
        id:
          type: string
          description: Auto-generated unique identifier
        name:
          type: string
          description: Name of the ticket holder or event
        date:
          type: string
          format: date
          description: Date of the event
        time:
          type: string
          description: Time of the event in HH:mm format
        movie:
          type: string
          description: ID of the movie
        schedule:
          type: string
          description: ID of the schedule
        screen:
          type: number
          description: Screen number
        seat:
          type: array
          items:
            type: string
          description: List of seat numbers
        seatType:
          type: string
          description: Type of seat (e.g., regular, VIP)
        price:
          type: number
          description: Price of the ticket
      example:
        id: 60c72b2f9e7d3c001f1d1d6f
        name: John Doe
        date: '2024-08-10'
        time: '19:30'
        movie: 60c72b2f9e7d3c001f1d1d6f
        schedule: 60c72b2f9e7d3c001f1d1d6f
        screen: 1
        seat:
          - A12
          - A13
        seatType: VIP
        price: 25
    UserRequest:
      type: object
      required:
        - name
        - email
        - phone
        - password
      properties:
        id:
          type: string
          description: Auto-generated unique identifier
        name:
          type: string
          description: Name of the user
        email:
          type: string
          description: Email of the user
        phone:
          type: string
          description: Phone number of the user
        password:
          type: string
          description: Password of the user
      example:
        name: John Doe
        email: john.doe@example.com
        phone: '1234567890'
        password: password
    UserResponse:
      type: object
      required:
        - name
        - email
        - phone
        - password
      properties:
        id:
          type: string
          description: Auto-generated unique identifier
        name:
          type: string
          description: Name of the user
        email:
          type: string
          description: Email of the user
        phone:
          type: string
          description: Phone number of the user
        password:
          type: string
          description: Hashed Password of the user
      example:
        name: John Doe
        email: john.doe@example.com
        phone: '1234567890'
        password: $2b$10$olHovuYBAwJm74.8Xe2Ku.hkdrB8F7PvJgHigNEO.PNcXOW1eSmze
  parameters:
...
    ticketId:
      name: id
      in: path
      description: ID of the ticket
      required: true
      schema:
        type: string
    userId:
      name: id
      in: path
      description: ID of the user
      required: true
      schema:
        type: string
...